LARA answer conversion script: reorganize fiels, rm unnecessary conversion variant
[#182525578]
