--- src/convert-lara-answers/global.d.ts
@@ -1,6 +1,0 @@
-declare module "*.scss";
-declare module "*.svg";
-declare module "shutterbug";
-declare module "iframe-phone";
-declare module "drawing-tool";
-declare module "react-dnd-preview";
