--- src/convert-lara-answers/index2.ts
@@ -1,262 +1,0 @@
-import { Firestore } from "@google-cloud/firestore";
-import { BigBatch } from "@qualdesk/firestore-big-batch";
-import { createTraverser } from '@firecode/admin';
-import deepEqual from "deep-equal";
-import * as fs from "fs";
-import { convertAnswer, getAnswerType, utcString } from "./utils";
-import { ILARAAnonymousAnswerReportHash, ILARAAnswerReportHash } from "./types";
-import { credentials, oldSourceKey, newSourceKey, resourceLinkId, batchedWrites, maxDocCount, startDate, endDate } from "./config.json";
-
-process.env.GOOGLE_APPLICATION_CREDENTIALS = credentials;
-if (!fs.existsSync(process.env.GOOGLE_APPLICATION_CREDENTIALS)) {
-  console.error("Missing or incorrect credentials - please create and download a credentials json file here: https://console.cloud.google.com/apis/credentials?project=report-service-pro&organizationId=264971417743");
-  process.exit(1);
-}
-
-const firestore = new Firestore();
-
-const logFile = `./log/info-${utcString()}.json`;
-const errorFile = `./log/error-${utcString()}.json`;
-
-const logLastLine = "]\n";
-
-const writeToFile = (file: string, content: string) => {
-  if (content === logLastLine) {
-    content = "\n" + content;
-  } else {
-    content = ",\n" + content;
-  }
-  fs.appendFileSync(file, content);
-};
-
-const getErrorHandler = (id: string) => (error: Error) => {
-  writeToFile(errorFile, JSON.stringify({
-    sourceId: id,
-    error: error.toString(),
-    stack: error.stack
-  }));
-};
-
-const log = (message: string) => {
-  writeToFile(logFile, message);
-};
-
-const getAllQuestions = (resource: any) => {
-  const result: any = [];
-
-  const activities = resource.type === "sequence" ? resource.children : [ resource ];
-
-  activities.forEach((activity: any) => {
-    activity.children.forEach((section: any) => {
-      section.children.forEach((page: any) => {
-        page.children.forEach((question: any) => {
-          result.push(question);
-        });
-      });
-    });
-  });
-
-  return result;
-};
-
-const generateQuestionMap = async () => {
-  process.stdout.write("Building question map: ");
-
-  try {
-    const data = fs.readFileSync('./question-map.json', {encoding:'utf8', flag:'r'});
-    const questionMapFromDisk = JSON.parse(data);
-    return { questionMap: questionMapFromDisk, processedActivitiesCount: -1 };
-  } catch(e) {
-    // continue;
-  }
-
-  let processedActivitiesCount = 0;
-
-  const resourcesToMigrate = firestore
-  .collection(`sources/${oldSourceKey}/resources`)
-  .where("migration_status", "==", "migrated")
-  // migration_test can be manually set in Firestore UI to limit conversion just to one activity.
-  // .where ("migration_test", "==", "true")
-  .where("created", ">=", startDate)
-  .where("created", "<=", endDate);
-
-  const resourcesTraverser = createTraverser(resourcesToMigrate, {
-    batchSize: 1000,
-    // This means we are prepared to hold batchSize * maxConcurrentBatchCount documents in memory.
-    // We sacrifice some memory to traverse faster.
-    maxConcurrentBatchCount: 10,
-  });
-
-  const questionMap: Record<string, any> = {};
-
-  await resourcesTraverser.traverse(async (resourceBatchDocs) => {
-    await Promise.all(resourceBatchDocs.map(async (resourceDoc) => {
-      const activityErrorHandler = getErrorHandler(resourceDoc.id);
-      try {
-        const resource = resourceDoc.data();
-        const questions = getAllQuestions(resource);
-
-        questions.forEach((question: any) => {
-          let safeQuestion = question;
-
-          // TODO: remove in the final version of the script. This is only necessary in the early tests when questions
-          // are not converted yet.
-          if (question.type === "multiple_choice" && !question.authored_state) {
-            // We're processing old MC question. It's impossible to convert its answer without authored
-            // state. Other question types can work without access to authored state.
-            safeQuestion = {
-              ...question,
-              authored_state: {
-                choices: [
-                  {
-                    id: 1,
-                    content: "a",
-                    correct: true
-                  },
-                  {
-                    id: 2,
-                    content: "b"
-                  },
-                ]
-              }
-            };
-          }
-
-          let sourceQuestionId;
-          if (question.legacy_id) {
-            sourceQuestionId = question.legacy_id;
-          } else {
-            // This is case of the managed interactives. They are not converted. We just need to move answers
-            // from one collection to another.
-            sourceQuestionId = question.id;
-          }
-
-          const questionKey = `${resource.url}-${sourceQuestionId}`;
-          // This might happen when one sequence has one activity added multiple times. But then all the questions
-          // should be identical.
-          if (questionMap[questionKey] && !deepEqual(safeQuestion, questionMap[questionKey])) {
-            writeToFile(errorFile, JSON.stringify({
-              type: "question-diff",
-              question1: safeQuestion,
-              question2: questionMap[questionKey],
-            }, null, 2));
-          }
-
-          questionMap[questionKey] = safeQuestion;
-        });
-
-        if (processedActivitiesCount % 100 === 0) {
-          process.stdout.write(".");
-        }
-        processedActivitiesCount += 1;
-
-      } catch (error) {
-        activityErrorHandler(error);
-      }
-    }));
-  });
-
-  fs.writeFileSync('./question-map.json', JSON.stringify(questionMap));
-
-  return { questionMap, processedActivitiesCount };
-};
-
-// Create a new client
-const executeScript = async () => {
-  // {} is added so the log file is always correct correct
-  fs.appendFileSync(logFile, `[\n{"logType": "info", "date": "${utcString()}"}`);
-  fs.appendFileSync(errorFile, `[\n{"logType": "error", "date": "${utcString()}"}`);
-
-  console.log("Answer conversion started");
-
-  const { questionMap, processedActivitiesCount } = await generateQuestionMap();
-
-  let copiedAnswersCount = 0;
-
-  const answersRef = firestore.collection(`sources/${oldSourceKey}/answers`);
-
-  const answersTraverser = createTraverser(answersRef, {
-    // 500 is the max batched write size, but we use BigBatch helper so any value can work here.
-    batchSize: 1000,
-    // This means we are prepared to hold batchSize * maxConcurrentBatchCount documents in memory.
-    // We sacrifice some memory to traverse faster.
-    maxConcurrentBatchCount: 10,
-    sleepTimeBetweenBatches: 0,
-    // TODO: remove limit in the final script.
-    maxDocCount
-  });
-
-  console.log("\nAnswer conversion: ");
-
-  const startTime = performance.now();
-
-  await answersTraverser.traverse(async (answerBatchDocs) => {
-
-    const answersUpdateBatch = batchedWrites ? new BigBatch({ firestore }) : null;
-
-    await Promise.all(answerBatchDocs.map(async (answerDoc) => {
-      const answerErrorHandler = getErrorHandler(answerDoc.id);
-      try {
-        const answer = answerDoc.data() as ILARAAnswerReportHash | ILARAAnonymousAnswerReportHash;
-        const answerType = getAnswerType(answer);
-        if (!answerType) {
-          throw new Error(`unknown answer type: ${answer.id}`);
-        }
-        if (answerType === "labbook") {
-          // Do nothing, labbooks don't require conversions. Activities will be marked as defunct.
-          return;
-        }
-
-        const question = questionMap[`${answer.resource_url}-${answer.question_id}`];
-        if (!question) {
-          // Nothing to do, question have not been converted.
-          return;
-        }
-
-        const convertedAnswer = convertAnswer(answerType, {
-          oldAnswer: answer,
-          newQuestion: question,
-          oldSourceKey,
-          newSourceKey,
-          additionalMetadata: {}
-        });
-
-        const answerRef = firestore.collection(`sources/${newSourceKey}/answers`).doc(convertedAnswer.id);
-
-        if (answersUpdateBatch) {
-          answersUpdateBatch.set(answerRef, convertedAnswer);
-        } else {
-          await answerRef.set(convertedAnswer);
-        }
-
-        copiedAnswersCount += 1;
-      } catch (error) {
-        answerErrorHandler(error);
-      }
-    }));
-
-    if (answersUpdateBatch) {
-      await answersUpdateBatch.commit();
-    }
-
-    process.stdout.write(".");
-  });
-
-
-  log(logLastLine);
-  writeToFile(errorFile, logLastLine);
-
-  const durationInS = (performance.now() - startTime) / 1000;
-  const activitySpeed = processedActivitiesCount / durationInS;
-  const answerSpeed = copiedAnswersCount / durationInS;
-
-  const used = process.memoryUsage().heapUsed / 1024 / 1024;
-
-  console.log(`\n\nThe script uses approximately ${used.toFixed(2)} MB of memory`);
-  console.log(`Elapsed time: ${durationInS.toFixed(2)} seconds`);
-  console.log(`Processed ${processedActivitiesCount} activities, ${activitySpeed.toFixed(2)} activities per second`);
-  console.log(`QuestionMap size: ${Object.keys(questionMap).length}`);
-  console.log(`Copied ${copiedAnswersCount} answers, ${answerSpeed.toFixed(2)} answers per second`);
-};
-
-executeScript();
