+++ src/convert-lara-answers/package.json
@@ -2,7 +2,7 @@
   "name": "convert-lara-answers",
   "version": "0.0.1",
   "description": "",
-  "main": "index.js",
+  "main": "src/convert-answers.ts",
   "jest": {
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
@@ -27,8 +27,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "start": "ts-node index.ts",
-    "start2": "ts-node index2.ts",
+    "start": "ts-node src/convert-answers.ts",
     "test": "jest",
     "start:prod": "npm run build && node dist/convert-lara-answers/index.js"
   },
