+++ src/convert-lara-answers/src/conversion-helpers/convert-open-response.ts
@@ -1,7 +1,7 @@
-import { IInteractiveState as IOpenResponseInteractiveState } from "../open-response/components/types";
+import { IInteractiveState as IOpenResponseInteractiveState } from "../../../open-response/components/types";
 import { convertAnswer } from "./convert-answer";
-import { ILARAAnonymousAnswerReportHash, ILARAAnswerReportHash, ILARAOpenResponseAnswerReportHash, IManagedInteractiveQuestion } from "./types";
-import { getReportState } from "./utils";
+import { ILARAAnonymousAnswerReportHash, ILARAAnswerReportHash, ILARAOpenResponseAnswerReportHash, IManagedInteractiveQuestion } from "../types";
+import { getReportState } from "../utils";
 
 export interface IConvertOpenResponseOptions {
   newQuestion: IManagedInteractiveQuestion;
