+++ src/convert-lara-answers/src/global.d.ts
@@ -1,0 +1,2 @@
+declare module "*.scss";
+declare module "*.svg";
