+++ src/convert-lara-answers/tsconfig.json
@@ -20,6 +20,6 @@
     ]
   },
   "include": [
-    "./index.ts"
+    "src/convert-answers.ts"
   ]
 }
